Allow custom-command to "install" itself
The comment in https://github.com/pypa/setuptools/issues/2591#issuecomment-791628028
reports that users wanting to use the custom-command plugin need to
explicitly add it using `setup.py`.

This commit tried to demonstrate how that can be done without requiring
the user to change their `setup.py`.

The "installation" of the new sub-command is done via the
`finalize_distribution_options` hook.